cleaned up figs 2, 3, 4
