+++ dev/2024_Optics_Express/figure_3.py
@@ -1,22 +1,28 @@
 # %% 
 # Imports
 # --------------------------------------------------------------------
-import os
 from pathlib import Path
-import sys
-import math
+
 import torch
 import torchvision
 import numpy as np
 import matplotlib.pyplot as plt
 
+import spyrit.core.meas as meas
+import spyrit.core.noise as noise
+import spyrit.core.prep as prep
 import spyrit.core.recon as recon
+import spyrit.core.nnet as nnet
+import spyrit.core.train as train
+import spyrit.misc.statistics as stats
 import spyrit.external.drunet as drunet
+
+import utility_dpgd as dpgd
+
 
 # %% 
 # General
 # --------------------------------------------------------------------
-
 # Experimental data
 image_folder = 'data/images/'       # images for simulated measurements
 model_folder = 'model/'             # reconstruction models
@@ -33,11 +39,11 @@ recon_folder_full.mkdir(parents=True, exist_ok=True)
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 print("Using device:", device)
 
+
 # %% 
 # Load images
 # --------------------------------------------------------------------
-import spyrit.misc.statistics as stats
-from spyrit.misc.disp import imagesc
+
 img_size = 128 # image size
 
 print("Loading image...")
@@ -64,7 +70,6 @@ c, h, w = x.shape
 print("Image shape:", x.shape)
 
 x_plot = x.view(-1, h, h).cpu().numpy()
-# imagesc(x_plot[0, :, :])
 # save image as original
 plt.imsave(recon_folder_full / 'original.png', x_plot[0, :, :], cmap='gray')
 
@@ -72,15 +77,9 @@ plt.imsave(recon_folder_full / 'original.png', x_plot[0, :, :], cmap='gray')
 # %% 
 # Simulate measurements for three image intensities
 # --------------------------------------------------------------------
+# Measurement parameters
 alpha_list = [2, 10, 50] # Poisson law parameter for noisy image acquisitions
 n_alpha = len(alpha_list)
-
-from spyrit.core.meas import HadamSplit
-from spyrit.core.noise import Poisson
-# from spyrit.misc.sampling import meas2img
-from spyrit.core.prep import SplitPoisson
-
-# Measurement parameters
 M = 128 * 128 // 4  # Number of measurements (here, 1/4 of the pixels)
 
 # Measurement and noise operators
@@ -88,9 +87,9 @@ Ord_rec = np.ones((img_size, img_size))
 Ord_rec[:,img_size//2:] = 0
 Ord_rec[img_size//2:,:] = 0
 
-meas_op = HadamSplit(M, h, torch.from_numpy(Ord_rec))
-noise_op = Poisson(meas_op, alpha_list[0])
-prep_op = SplitPoisson(2, meas_op)
+meas_op = meas.HadamSplit(M, h, torch.from_numpy(Ord_rec))
+noise_op = noise.Poisson(meas_op, alpha_list[0])
+prep_op = prep.SplitPoisson(2, meas_op)
 
 # Vectorized image
 x = x.view(1, h * w)
@@ -106,13 +105,12 @@ for ii, alpha in enumerate(alpha_list):
 y = y.to(device)
 
 
+
 # %% 
 # Pinv
 # ====================================================================
-from spyrit.core.recon import PinvNet
-
 # Init
-pinv = PinvNet(noise_op, prep_op)
+pinv = recon.PinvNet(noise_op, prep_op)
 
 # Use GPU if available
 pinv = pinv.to(device)
@@ -134,19 +132,14 @@ with torch.no_grad():
 # %% 
 # Pinv-Net
 # ====================================================================
-from spyrit.core.recon import PinvNet
-from spyrit.core.nnet import Unet
-from spyrit.core.train import load_net
-
-# retrained is same as original
 model_name = 'pinv-net_unet_imagenet_N0_10_m_hadam-split_N_128_M_4096_epo_30_lr_0.001_sss_10_sdr_0.5_bs_512_reg_1e-07_retrained_light.pth'
 
 # Init
-pinvnet = PinvNet(noise_op, prep_op, Unet())
+pinvnet = recon.PinvNet(noise_op, prep_op, nnet.Unet())
 pinvnet.eval() 
 
 # Load net and use GPU if available
-load_net(model_folder_full / model_name, pinvnet, device, False)
+train.load_net(model_folder_full / model_name, pinvnet, device, False)
 pinvnet = pinvnet.to(device)
 
 # Reconstruct
@@ -166,10 +159,6 @@ with torch.no_grad():
 # %% 
 # DC-Net
 # ====================================================================
-from spyrit.core.recon import DCNet
-from spyrit.core.nnet import Unet
-from spyrit.core.train import load_net
-
 model_name = 'dc-net_unet_imagenet_rect_N0_10_N_128_M_4096_epo_30_lr_0.001_sss_10_sdr_0.5_bs_256_reg_1e-07_light.pth'
 cov_name = stat_folder_full / 'Cov_8_{}x{}.npy'.format(img_size, img_size)
 
@@ -177,12 +166,12 @@ cov_name = stat_folder_full / 'Cov_8_{}x{}.npy'.format(img_size, img_size)
 Cov = torch.from_numpy(np.load(cov_name)) 
 
 # Init
-denoi = Unet()         # torch.nn.Identity()
-dcnet = DCNet(noise_op, prep_op, Cov, denoi)
+denoi = nnet.Unet()         # torch.nn.Identity()
+dcnet = recon.DCNet(noise_op, prep_op, Cov, denoi)
 dcnet.eval() 
 
 # Load net and use GPU if available
-load_net(model_folder_full / model_name, dcnet, device, False)
+train.load_net(model_folder_full / model_name, dcnet, device, False)
 dcnet = dcnet.to(device)
 
 # Reconstruct
@@ -202,16 +191,15 @@ with torch.no_grad():
 # %% 
 # LPGD
 # ====================================================================
-
 model_name = "lpgd_unet_imagenet_N0_10_m_hadam-split_N_128_M_4096_epo_30_lr_0.001_sss_10_sdr_0.5_bs_128_reg_1e-07_uit_3_sdec0-9_light.pth"
 
 # Initialize network
-denoi = Unet()
+denoi = nnet.Unet()
 lpgd = recon.LearnedPGD(noise_op, prep_op, denoi, step_decay=0.9)
 lpgd.eval()
 
 # load net and use GPU if available
-load_net(model_folder_full / model_name, lpgd, device, False)
+train.load_net(model_folder_full / model_name, lpgd, device, False)
 lpgd = lpgd.to(device)
 
 # Reconstruct and save
@@ -232,7 +220,6 @@ with torch.no_grad():
 # %% 
 # Pinv - PnP
 # ====================================================================
-
 model_name = "drunet_gray.pth"
 noise_levels = [115, 45, 20] # noise levels from 0 to 255 for each alpha
 
@@ -267,13 +254,13 @@ with torch.no_grad():
             cmap='gray')
 
 
-# %% DPGD-PnP
-from utility_dpgd import load_model, DualPGD
-
+# %% 
+# DPGD-PnP
+# ====================================================================
 # load denoiser
 n_channel, n_feature, n_layer = 1, 100, 20
 model_name = 'DFBNet_l1_patchsize=50_varnoise0.1_feat_100_layers_20.pth'
-denoi = load_model(pth = (model_folder_full / model_name).as_posix(), 
+denoi = dpgd.load_model(pth = (model_folder_full / model_name).as_posix(), 
                     n_ch = n_channel, 
                     features = n_feature, 
                     num_of_layers = n_layer)
@@ -288,7 +275,7 @@ mu_list = [6000, 3500, 1500]
 crit_norm = 1e-4
 
 # Init 
-dpgdnet = DualPGD(noise_op, prep_op, denoi, gamma, mu_list[0], max_iter, crit_norm)
+dpgdnet = dpgd.DualPGD(noise_op, prep_op, denoi, gamma, mu_list[0], max_iter, crit_norm)
 x_dpgd = torch.zeros(1, 1, img_size, img_size)
 
 with torch.no_grad():
