+++ dev/2024_Optics_Express/figure_4.py
@@ -1,15 +1,24 @@
-# %% Imports
+# %% 
+# Imports
 # --------------------------------------------------------------------
-import json
 import ast
+import json
 from pathlib import Path
+
 import torch
 import numpy as np
 import matplotlib.pyplot as plt
 
+import spyrit.core.meas as meas
+import spyrit.core.noise as noise
+import spyrit.core.prep as prep
 import spyrit.core.recon as recon
+import spyrit.core.nnet as nnet
+import spyrit.core.train as train
+import spyrit.misc.statistics as stats
 import spyrit.external.drunet as drunet
-import spyrit.misc.sampling as samp
+
+import utility_dpgd as dpgd
 
 
 # %% General
@@ -53,10 +62,7 @@ subsampling_factor = 2 # measure an equivalent image of size img_size // 2
 # %% Measurement (and reconstruction) operators
 # --------------------------------------------------------------------
 
-from spyrit.core.meas import HadamSplit
-from spyrit.core.noise import Poisson
-# from spyrit.misc.sampling import meas2img
-from spyrit.core.prep import SplitPoisson
+
 
 # number of measurements
 M = (img_size // subsampling_factor)**2
@@ -66,9 +72,9 @@ Ord_rec = np.ones((img_size, img_size))
 Ord_rec[:,img_size//2:] = 0
 Ord_rec[img_size//2:,:] = 0
 
-meas_op = HadamSplit(M, img_size, torch.from_numpy(Ord_rec))
-noise_op = Poisson(meas_op, 2) # parameter alpha is unimportant here
-prep_op = SplitPoisson(2, meas_op) # same here
+meas_op = meas.HadamSplit(M, img_size, torch.from_numpy(Ord_rec))
+noise_op = noise.Poisson(meas_op, 2) # parameter alpha is unimportant here
+prep_op = prep.SplitPoisson(2, meas_op) # same here
 
 
 # %% Load experimental data and prepare it for reconstruction
@@ -126,10 +132,8 @@ for i in range(n_meas):
 
 # %% Pinv
 # ====================================================================
-from spyrit.core.recon import PinvNet
-
 # Init
-pinv = PinvNet(noise_op, prep_op)
+pinv = recon.PinvNet(noise_op, prep_op)
 
 # Use GPU if available
 pinv = pinv.to(device)
@@ -145,23 +149,19 @@ with torch.no_grad():
         filename = f'pinv_{savenames[ii]}.png'
         full_path = recon_folder_full / filename
         plt.imsave(full_path, x_pinv[0, 0, :, :].cpu().detach().numpy(), 
-            cmap='gray') #
+            cmap='gray')
 
 
 # %% Pinv-Net
 # ====================================================================
-from spyrit.core.nnet import Unet
-from spyrit.core.train import load_net
-
-# retrained is same as original
 model_name = 'pinv-net_unet_imagenet_N0_10_m_hadam-split_N_128_M_4096_epo_30_lr_0.001_sss_10_sdr_0.5_bs_512_reg_1e-07_retrained_light.pth'
 
 # Init
-pinvnet = PinvNet(noise_op, prep_op, Unet())
+pinvnet = recon.PinvNet(noise_op, prep_op, nnet.Unet())
 pinvnet.eval() 
 
 # Load net and use GPU if available
-load_net(model_folder_full / model_name, pinvnet, device, False)
+train.load_net(model_folder_full / model_name, pinvnet, device, False)
 pinvnet = pinvnet.to(device)
 
 # Reconstruct
@@ -175,15 +175,11 @@ with torch.no_grad():
         filename = f'pinvnet_{savenames[ii]}.png'
         full_path = recon_folder_full / filename
         plt.imsave(full_path, x_pinvnet[0, 0, :, :].cpu().detach().numpy(), 
-            cmap='gray') #
+            cmap='gray')
 
 
 # %% DC-Net
 # ====================================================================
-from spyrit.core.recon import DCNet
-from spyrit.core.nnet import Unet
-from spyrit.core.train import load_net
-
 model_name = 'dc-net_unet_imagenet_rect_N0_10_N_128_M_4096_epo_30_lr_0.001_sss_10_sdr_0.5_bs_256_reg_1e-07_light.pth'
 cov_name = stat_folder_full / 'Cov_8_{}x{}.npy'.format(img_size, img_size)
 
@@ -191,12 +187,12 @@ cov_name = stat_folder_full / 'Cov_8_{}x{}.npy'.format(img_size, img_size)
 Cov = torch.from_numpy(np.load(cov_name)) 
 
 # Init
-denoi = Unet()         # torch.nn.Identity()
-dcnet = DCNet(noise_op, prep_op, Cov, denoi)
+denoi = nnet.Unet()         # torch.nn.Identity()
+dcnet = recon.DCNet(noise_op, prep_op, Cov, denoi)
 dcnet.eval() 
 
 # Load net and use GPU if available
-load_net(model_folder_full / model_name, dcnet, device, False)
+train.load_net(model_folder_full / model_name, dcnet, device, False)
 dcnet = dcnet.to(device)
 
 # Reconstruct
@@ -210,21 +206,20 @@ with torch.no_grad():
         filename = f'dcnet_{savenames[ii]}.png'
         full_path = recon_folder_full / filename
         plt.imsave(full_path, x_dcnet[0, 0, :, :].cpu().detach().numpy(), 
-            cmap='gray') #
+            cmap='gray')
 
 
 # %% LPGD
 # ====================================================================
-
 model_name = "lpgd_unet_imagenet_N0_10_m_hadam-split_N_128_M_4096_epo_30_lr_0.001_sss_10_sdr_0.5_bs_128_reg_1e-07_uit_3_sdec0-9_light.pth"
 
 # Initialize network
-denoi = Unet()
+denoi = nnet.Unet()
 lpgd = recon.LearnedPGD(noise_op, prep_op, denoi, step_decay=0.9)
 lpgd.eval()
 
 # load net and use GPU if available
-load_net(model_folder_full / model_name, lpgd, device, False)
+train.load_net(model_folder_full / model_name, lpgd, device, False)
 lpgd = lpgd.to(device)
 
 # Reconstruct
@@ -243,7 +238,6 @@ with torch.no_grad():
 
 # %% Pinv - PnP
 # ====================================================================
-
 model_name = "drunet_gray.pth"
 noise_levels = [55, 35] # noise levels from 0 to 255 for each alpha
 
@@ -276,12 +270,10 @@ with torch.no_grad():
 
 # %% DPGD-PnP
 # ====================================================================
-from utility_dpgd import load_model, DualPGD
-
 # load denoiser
 n_channel, n_feature, n_layer = 1, 100, 20
 model_name = 'DFBNet_l1_patchsize=50_varnoise0.1_feat_100_layers_20.pth'
-denoi = load_model(pth = (model_folder_full / model_name).as_posix(), 
+denoi = dpgd.load_model(pth = (model_folder_full / model_name).as_posix(), 
                     n_ch = n_channel, 
                     features = n_feature, 
                     num_of_layers = n_layer)
@@ -297,7 +289,7 @@ mu_list = [4000, 4000]
 crit_norm = 1e-4
 
 # Init 
-dpgdnet = DualPGD(noise_op, prep_op, denoi, gamma, mu_list[0], max_iter, crit_norm)
+dpgdnet = dpgd.DualPGD(noise_op, prep_op, denoi, gamma, mu_list[0], max_iter, crit_norm)
 x_dpgd = torch.zeros(1, 1, img_size, img_size)
 
 with torch.no_grad():
